add more items to the querires and hardening
